// src/generated/models/Sd_requisitioncommentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Sd_requisitioncommentssd_kind = {
  100000000: 'Comment',
  100000001: 'Clarification',
  100000002: 'BudgetNote',
  100000003: 'System'
} as const;
export type Sd_requisitioncommentssd_kind = keyof typeof Sd_requisitioncommentssd_kind;
export const Sd_requisitioncommentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Sd_requisitioncommentsstatecode = keyof typeof Sd_requisitioncommentsstatecode;
export const Sd_requisitioncommentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Sd_requisitioncommentsstatuscode = keyof typeof Sd_requisitioncommentsstatuscode;

export interface Sd_requisitioncommentsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  sd_authorname?: string;
  sd_authorpersonnelnumber?: string;
  sd_body: string;
  sd_kind?: Sd_requisitioncommentssd_kind;
  sd_linenumber?: number;
  sd_name?: string;
  sd_parentcommentid?: string;
  sd_requisitioncommentid: string;
  sd_requisitionnumber: string;
  statecode: Sd_requisitioncommentsstatecode;
  statuscode?: Sd_requisitioncommentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Sd_requisitioncomments extends Sd_requisitioncommentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  sd_attachment?: string;
  sd_attachment_name?: string;
  sd_kindname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Sd_requisitioncommentsUploadColumnName = 'sd_attachment';